sspi & ffi: add doc comments about creds and sspi tables; small refactoring
